Refactor: 위젯의 위치 변경
